+++ packages/host/src/components/Navigation/ReplenishmentNav.tsx
@@ -1,33 +1,15 @@
-import React, { FC, useEffect } from 'react';
-import { useMatch } from 'react-router-dom';
+import React, { FC } from 'react';
+
 import {
-  TruckIcon,
+  SuppliersIcon,
   Collapse,
   List,
-  useDrawer,
   useTranslation,
   RouteBuilder,
   NavLink,
 } from '@openmsupply-client/common';
 import { AppRoute } from '@openmsupply-client/config';
-
-const matchPath = (key: string, path: string) =>
-  `/${key.replace(/^\//, '')}/`.startsWith(path.replace(/\*$/, ''));
-
-const useNestedNav = (path: string) => {
-  const { hoverActive, isOpen } = useDrawer();
-  const match = useMatch(path);
-  const [expanded, setExpanded] = React.useState(false);
-  const hovered = Object.keys(hoverActive).some(
-    key => matchPath(key, path) && hoverActive[key]
-  );
-
-  useEffect(() => {
-    setExpanded(!!match);
-  }, [match]);
-
-  return { isActive: isOpen && (expanded || hovered) };
-};
+import { useNestedNav } from './useNestedNav';
 
 export const ReplenishmentNav: FC = () => {
   const { isActive } = useNestedNav(
@@ -40,7 +22,7 @@ export const ReplenishmentNav: FC = () => {
       <NavLink
         end={false}
         to={AppRoute.Replenishment}
-        icon={<TruckIcon color="primary" fontSize="small" />}
+        icon={<SuppliersIcon color="primary" fontSize="small" />}
         expandOnHover
         text={t('replenishment')}
       />
